refactor: SplitManager 삭제후 CalculationSegmentProcessor 로 기능 편입
