+++ src/main/java/calculator/Application.java
@@ -5,7 +5,6 @@ import calculator.domain.mainprocess.NumCalculator;
 import calculator.domain.preprocess.DelimeterSegmentProcessor;
 import calculator.domain.preprocess.CalculationSegmentProcessor;
 import calculator.domain.preprocess.Preprocessor;
-import calculator.domain.preprocess.SplitManager;
 import calculator.view.ViewManager;
 
 public class Application {
@@ -14,8 +13,7 @@ public class Application {
                 new NumCalculator(),
                 new Preprocessor(
                         new DelimeterSegmentProcessor(),
-                        new CalculationSegmentProcessor(),
-                        new SplitManager()),
+                        new CalculationSegmentProcessor()),
                 new ViewManager());
 
         calculatorController.runCalculator();
