+++ src/main/java/calculator/domain/preprocess/CalculationSegmentProcessor.java
@@ -1,5 +1,7 @@
 package calculator.domain.preprocess;
 
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 import static calculator.utils.Util.isInteger;
@@ -18,5 +20,19 @@ public class CalculationSegmentProcessor {
                 throw new IllegalArgumentException("구분자 이외의 문자가 있습니다.");
         }
     }
+
+    private List<String> discardBlankSpace(List<String> input) {
+        input.removeIf(String::isBlank);
+        return input;
+    }
+
+    public List<String> extractNumbers(List<String> delimeterList, String input){
+
+        String regularExpression = "["+String.join("",delimeterList)+"]";
+        String[] extractedNumbers = input.split(regularExpression);
+        List<String> extractedNumbersList = new ArrayList<>(Arrays.asList(extractedNumbers));
+        return  discardBlankSpace(extractedNumbersList);
+    }
+
 
 }
