--- src/main/java/calculator/domain/preprocess/SplitManager.java
@@ -1,22 +1,0 @@
-package calculator.domain.preprocess;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
-public class SplitManager {
-
-    private List<String> discardBlankSpace(List<String> input) {
-        input.removeIf(String::isBlank);
-        return input;
-    }
-
-    public List<String> extractNumbers(List<String> delimeterList, String input){
-
-        String regularExpression = "["+String.join("",delimeterList)+"]";
-        String[] extractedNumbers = input.split(regularExpression);
-        List<String> extractedNumbersList = new ArrayList<>(Arrays.asList(extractedNumbers));
-        return  discardBlankSpace(extractedNumbersList);
-    }
-
-}
